0. Float like a butterfly, sting like a bee

+++ 0x0B-malloc_free/0-create_array.c
@@ -21,7 +21,7 @@ char *create_array(unsigned int size, char c)
 		return (NULL);
 	}
 
-	while (i < size)
+	while (i < (int)size)
 	{
 		*(arr + i) = c;
 		i++;
